refactor(db): unify RDB package with DB package
