+++ GNUmakefile
@@ -22,7 +22,7 @@
 	diff-server-rdb-redis
 
 SRCS = $(shell git ls-files '*.go')
-PKGS =  ./commands ./config ./db ./db/rdb ./fetcher ./models ./util
+PKGS =  ./commands ./config ./db ./fetcher ./models ./util
 VERSION := $(shell git describe --tags --abbrev=0)
 REVISION := $(shell git rev-parse --short HEAD)
 LDFLAGS := -X 'github.com/kotakanbe/goval-dictionary/config.Version=$(VERSION)' \
