+++ commands/fetch-ubuntu.go
@@ -37,7 +37,7 @@ func fetchUbuntu(cmd *cobra.Command, args []string) (err error) {
 		return xerrors.New("Failed to fetch ubuntu command. err: specify versions to fetch")
 	}
 
-	driver, locked, err := db.NewDB(c.Ubuntu, viper.GetString("dbtype"), viper.GetString("dbpath"), viper.GetBool("debug-sql"))
+	driver, locked, err := db.NewDB(viper.GetString("dbtype"), viper.GetString("dbpath"), viper.GetBool("debug-sql"))
 	if err != nil {
 		if locked {
 			log15.Error("Failed to open DB. Close DB connection before fetching", "err", err)
@@ -103,7 +103,7 @@ func fetchUbuntu(cmd *cobra.Command, args []string) (err error) {
 			FileName:  ss[len(ss)-1],
 		}
 
-		if err := driver.InsertOval(c.Ubuntu, &root, fmeta); err != nil {
+		if err := driver.InsertOval(&root, fmeta); err != nil {
 			log15.Error("Failed to insert OVAL", "err", err)
 			return err
 		}
